Django form with database and Bootstrap
Simple example how to create form that will take data for track of
chemicals, store them in database and display them. Item of list can be
deleted by clicking on trash bin.  If current chemical is due to expire
in 30 days, such item will be displayed in warning red color. To test
it, copy all files to yor computer, run "python manage.py runserver" and
then in browser type "localhost: 8000/assets/chemicals or any of your
ports or type 127.0.0.1 instead of localhost. Menu is currently not
working and is placed as mockup.